requirements.txt
opencv-python == 4.5.4.60
numpy == 1.23.4
Pillow == 9.2.0
webcolors==1.10